requirements.txt
Django==4.1.5
djangorestframework==3.14.0
psycopg2==2.9.5
django-phonenumber-field == 7.0.2
django-phonenumbers == 1.0.1
djangorestframework-simplejwt == 5.2.2
celery[redis]==5.2.7
flower==1.2.0
celery_singleton == 0.3.1
